api/requirements.txt
fastapi 
uvicorn
langchain-deepseek
langchain-tavily
langgraph
langchain
python-dotenv
langchain-core
langchain-huggingface
qdrant-client
langchain-qdrant
pydantic
langchain-text-splitters
pdf2image
pytesseract
python-multipart